sql: migrate addRender() and checkRenderStar() to a new file.
